+++ CHANGELOG.md
@@ -4,6 +4,7 @@
    listening, code looks better now
    * Fixed setup inventory didn't work
    * Using new Scoreboard library, whole game will perform now much better without lags
+   * Now language.yml file is cached on load so it will drastically reduce lag of plugin
    > Pre 4 (11.01.2019-06.02.2019)
    * Added useSSL=false argument in mysql.yml address to avoid mysql warnings in console (not harmful)
    * Added %MAX_PLAYERS% placeholder in scoreboard